+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "prettier-vscode" extension will be documented in thi
 
 <!-- Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file. -->
 
+## [3.19.0]
+
+- Notifications only show when running formatter
+- Removed logs for status bar events
+
 ## [3.18.0]
 
 - Fixed a bug with the prettier output open command.